Авторизация и админка

+++ MainApp/views.py
@@ -1,10 +1,10 @@
-from django.http import Http404
+from django.http import Http404, HttpResponseForbidden
 from django.shortcuts import render, redirect
-
+from django.contrib import auth
 from MainApp.models import Snippet
-from MainApp.forms import SnippetForm
-
+from MainApp.forms import SnippetForm, UserRegistrationForm
 from django.core.exceptions import ObjectDoesNotExist
+from django.contrib.auth.decorators import login_required
 
 
 def index_page(request):
@@ -12,6 +12,18 @@ def index_page(request):
     return render(request, 'pages/index.html', context)
 
 
+@login_required
+def my_snippets(request):
+    snippets = Snippet.objects.filter(user=request.user)
+    context = {
+        'pagename': 'Мои сниппеты',
+        "snippets": snippets,
+        "count": snippets.count()
+        }
+    return render(request, 'pages/view_snippets.html', context)
+
+
+@login_required
 def add_snippet_page(request):
     # Хотим получить чистую форму для заполнения полей
     if request.method == "GET":
@@ -25,16 +37,20 @@ def add_snippet_page(request):
     if request.method == "POST":
         form = SnippetForm(request.POST)
         if form.is_valid():
-            form.save()
+            snippet = form.save(commit=False)
+            if request.user.is_authenticated:
+                snippet.user = request.user
+                snippet.save()
             return redirect("snippets_list")
         return render(request,'add_snippet.html', {'form': form})
 
 
 def snippets_page(request):
-    snippets = Snippet.objects.all()
+    snippets = Snippet.objects.filter(public=True)
     context = {
         'pagename': 'Просмотр сниппетов',
-        "snippets": snippets
+        "snippets": snippets,
+        'count': snippets.count()
         }
     return render(request, 'pages/view_snippets.html', context)
 
@@ -53,16 +69,20 @@ def snippet_detail(request, snippet_id):
         return render(request, 'pages/snippet_detail.html', context)
 
 
+@login_required
 def snippet_delete(request, snippet_id):
     try:
         snippet = Snippet.objects.get(id=snippet_id)
     except ObjectDoesNotExist:
         raise Http404
     else:
-        snippet.delete()
-        return redirect("snippets_list")
+        if snippet.user == request.user:
+            snippet.delete()
+            return redirect("snippets_list")
+        return HttpResponseForbidden("It's others snippets")
 
 
+@login_required
 def snippet_edit(request, snippet_id):
     try:
         snippet = Snippet.objects.get(id=snippet_id)
@@ -82,8 +102,47 @@ def snippet_edit(request, snippet_id):
         snippet.lang = data_form['lang']
         snippet.creation_date = data_form['creation_date']
         snippet.code = data_form['code']
+        snippet.public = data_form.get('public', False)
         snippet.save()
         return redirect('snippets_list')
+
+
+def login(request):
+   if request.method == 'POST':
+       username = request.POST.get("username")
+       password = request.POST.get("password")
+       user = auth.authenticate(request, username=username, password=password)
+       if user is not None:
+           auth.login(request, user)
+       else:
+           context = {
+               "pagename": "PythonBin",
+               "errors": ['wrong username or password']
+           }
+           return render(request, 'pages/index.html', context)
+   return redirect('home')
+
+
+def logout(request):
+    auth.logout(request)
+    return redirect('home')
+
+
+def create_user(request):
+    context = {'pagename': 'Регистрация пользователя'}
+    # Хотим получить чистую форму для заполнения полей
+    if request.method == "GET":
+        form = UserRegistrationForm
+        context['form'] = form
+        return render(request, 'pages/registration.html', context)
+    # Хотим создать нового пользователя на основе данных от формы
+    if request.method == "POST":
+        form = UserRegistrationForm(request.POST)
+        if form.is_valid():
+            form.save()
+            return redirect("home")
+        context['form'] = form
+        return render(request,'pages/registration.html', context)
 
 
 # def create_snippet(request):
